fix(cli): rotation size calculation

+++ src/Servy.CLI/Commands/InstallServiceCommand.cs
@@ -67,7 +67,7 @@ namespace Servy.CLI.Commands
                 var recoveryAction = ParseEnumOption(opts.RecoveryAction, RecoveryAction.RestartService);
 
                 // Parse numeric options
-                int rotationSize = (int.TryParse(opts.RotationSize, out var rot) ? rot : AppConfig.DefaultRotationSize) * 1024 * 1024;
+                ulong rotationSize = (ulong)(int.TryParse(opts.RotationSize, out var rot) ? rot : AppConfig.DefaultRotationSize) * 1024 * 1024;
                 int heartbeatInterval = int.TryParse(opts.HeartbeatInterval, out var hb) ? hb : AppConfig.DefaultHeartbeatInterval;
                 int maxFailedChecks = int.TryParse(opts.MaxFailedChecks, out var mf) ? mf : AppConfig.DefaultMaxFailedChecks;
                 int maxRestartAttempts = int.TryParse(opts.MaxRestartAttempts, out var mr) ? mr : AppConfig.DefaultMaxRestartAttempts;
@@ -87,7 +87,7 @@ namespace Servy.CLI.Commands
                     opts.StdoutPath,
                     opts.StderrPath,
                     opts.EnableRotation,
-                    (ulong)rotationSize,
+                    rotationSize,
                     opts.EnableHealthMonitoring,
                     heartbeatInterval,
                     maxFailedChecks,
